# backend/src/test/resources/openapi/api-docs.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "Mail Backend API",
    "description" : "REST API for the Tauri desktop mail client — accounts, messages, drafts, folders.",
    "license" : {
      "name" : "Proprietary"
    },
    "version" : "1.0.0"
  },
  "security" : [ {
    "ApiKeyAuth" : [ ]
  } ],
...
  "paths" : {
...
    "/api/v1/accounts/{accountId}/drafts" : {
      "get" : {
        "tags" : [ "Drafts" ],
        "summary" : "List drafts",
        "description" : "Returns a paginated list of drafts saved in the Drafts folder for the given account.",
        "operationId" : "listDrafts",
        "parameters" : [ {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          }
        }, {
          "name" : "page",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 0
          }
        }, {
          "name" : "size",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "minimum" : 1
          }
        } ],
        "responses" : {
          "503" : {
            "description" : "IMAP/SMTP server unavailable (MAIL_CONNECTION_ERROR).",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/PagedResponseMailSummaryResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      },
      "post" : {
        "tags" : [ "Drafts" ],
        "summary" : "Save draft (async)",
        "description" : "Asynchronously saves the draft into the Drafts folder. With the replaces={stableId} parameter, the old revision is deleted after a successful save. Returns 202 Accepted immediately — the draft's stableId becomes available after the next sync.",
        "operationId" : "saveDraft",
        "parameters" : [ {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          }
        }, {
          "name" : "replaces",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "string",
            "maxLength" : 128,
            "minLength" : 0
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/DraftRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "503" : {
            "description" : "IMAP/SMTP server unavailable (MAIL_CONNECTION_ERROR).",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "202" : {
            "description" : "Draft save request accepted; processing continues asynchronously."
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "400" : {
            "description" : "Invalid input (validation error).",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/{accountId}/drafts/{stableId}/send" : {
      "post" : {
        "tags" : [ "Drafts" ],
        "summary" : "Send draft (async)",
        "description" : "Asynchronously sends an existing draft over SMTP: the original MIME is fetched from the Drafts folder (preserving Message-ID, threading headers and attachments), sent, appended to Sent and then hard-deleted from Drafts. Returns 202 Accepted with a sendId — the client tracks the outcome via the send_completed / send_failed notification stream event.",
        "operationId" : "sendDraft",
        "parameters" : [ {
          "name" : "accountId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64",
            "exclusiveMinimum" : 0
          }
        }, {
          "name" : "stableId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "maxLength" : 128,
            "minLength" : 0
          }
        } ],
        "responses" : {
          "503" : {
            "description" : "IMAP/SMTP server unavailable (MAIL_CONNECTION_ERROR).",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "202" : {
            "description" : "Send request accepted; outcome is delivered via the notification stream.",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/SendAcceptedResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "Missing or invalid X-API-KEY.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          },
          "500" : {
            "description" : "Internal server error.",
            "content" : {
              "application/problem+json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ProblemDetail"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "ProblemDetail" : {
        "description" : "RFC 9457 problem+json — format of all error responses.",
        "properties" : {
          "type" : {
            "type" : "string",
            "format" : "uri",
            "description" : "URI identifier of the error type (e.g. .../errors/account_not_found)."
          },
          "title" : {
            "type" : "string",
            "description" : "Short human-readable description of the error class."
          },
          "status" : {
            "description" : "HTTP status code."
          },
          "detail" : {
            "type" : "string",
            "description" : "Error detail intended for the user."
          },
          "instance" : {
            "type" : "string",
            "format" : "uri",
            "description" : "URI of the specific occurrence (typically the request URI)."
          },
          "errorCode" : {
            "type" : "string",
            "description" : "Machine-readable error code (e.g. ACCOUNT_NOT_FOUND, VALIDATION_ERROR)."
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time",
            "description" : "Time when the error occurred (ISO-8601)."
          }
        }
      },
...
      "AttachmentRequest" : {
        "type" : "object",
        "properties" : {
          "fileName" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 0
          },
          "contentType" : {
            "type" : "string",
            "maxLength" : 255,
            "minLength" : 0
          },
          "base64Data" : {
            "type" : "string",
            "maxLength" : 73400320,
            "minLength" : 0
          }
        },
        "required" : [ "base64Data", "contentType", "fileName" ]
      },
...
      "SendAcceptedResponse" : {
        "type" : "object",
        "properties" : {
          "sendId" : {
            "type" : "string",
            "description" : "Correlation id for tracking the async send outcome over the notification stream.",
            "example" : "b3f1c2de-4a5b-6c7d-8e9f-0a1b2c3d4e5f"
          }
        }
      },
...
      "DraftRequest" : {
        "type" : "object",
        "properties" : {
          "to" : {
            "type" : "string"
          },
          "cc" : {
            "type" : "string"
          },
          "bcc" : {
            "type" : "string"
          },
          "subject" : {
            "type" : "string",
            "maxLength" : 500,
            "minLength" : 0
          },
          "body" : {
            "type" : "string"
          },
          "attachments" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/AttachmentRequest"
            }
          },
          "inReplyTo" : {
            "type" : "string"
          },
          "references" : {
            "type" : "string"
          }
        }
      },
...
      "MailSummaryResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "stableId" : {
            "type" : "string",
            "description" : "Stable message identifier for the REST API. The client uses it for message detail and message actions."
          },
          "folderName" : {
            "type" : "string"
          },
          "subject" : {
            "type" : "string"
          },
          "sender" : {
            "type" : "string"
          },
          "recipientsTo" : {
            "type" : "string",
            "description" : "Primary recipients (To). Useful for Drafts/Sent listings where the sender is the account owner."
          },
          "receivedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "seen" : {
            "type" : "boolean"
          },
          "flagged" : {
            "type" : "boolean"
          },
          "answered" : {
            "type" : "boolean"
          },
          "hasAttachments" : {
            "type" : "boolean"
          },
          "threadId" : {
            "type" : "string",
            "description" : "Conversation identifier — every message of the same thread shares it. Null only for messages persisted before the V2 backfill completes (see /api/internal/threading/recompute). Frontends MAY group rows by this id; the V0.1.0 desktop client ignores it."
          }
        }
      },
...
      "PagedResponseMailSummaryResponse" : {
        "type" : "object",
        "properties" : {
          "content" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/MailSummaryResponse"
            }
          },
          "page" : {
            "type" : "integer",
            "format" : "int32"
          },
          "size" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalPages" : {
            "type" : "integer",
            "format" : "int32"
          },
          "totalElements" : {
            "type" : "integer",
            "format" : "int64"
          },
          "first" : {
            "type" : "boolean"
          },
          "last" : {
            "type" : "boolean"
          }
        }
      },
...
    },
    "securitySchemes" : {
      "ApiKeyAuth" : {
        "type" : "apiKey",
        "description" : "Internal API key that the desktop client loads from session.json and sends in the X-API-KEY header.",
        "name" : "X-API-KEY",
        "in" : "header"
      }
    }
  }
}